+++ src/app/providers/RouterProvider.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from 'react';
 import { createBrowserRouter, RouterProvider as RRRouterProvider } from 'react-router-dom';
 import { appRoutes } from '@/app/routes';
-import { Spinner } from '@/shared/ui/atoms/spinner';
+import { SpinnerOverlay } from '@/shared/ui/elements/SpinnerOverlay';
 import { NavigationService } from '@/shared/lib/router/navigation';
 
 const router = createBrowserRouter(appRoutes, {
@@ -14,7 +14,7 @@ NavigationService.setNavigate(router.navigate.bind(router));
 
 export function RouterProvider() {
   return (
-    <Suspense fallback={<Spinner className="size-10 animate-spin" />}>
+    <Suspense fallback={<SpinnerOverlay className="h-screen" />}>
       <RRRouterProvider
         router={router}
         future={{